[CodeCompletion] Split CodeCompletionString into its own file
